requirements.txt
asgiref==3.9.1
attrs==25.3.0
coverage==7.10.3
dataclasses==0.6
Django==5.2.5
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djongo==1.2.31
dnspython==2.7.0
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pymongo==4.14.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
django-cors-headers==4.3.1
Pillow==10.4.0
django-filter==24.3